Cleaned up Teams implementation
