+++ src/sdk/PnP.Core.Test/Base/UpdateTests.cs
@@ -5,6 +5,7 @@ using PnP.Core.Model.Teams;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using PnP.Core.Model.Teams.Public.Enums;
 
 namespace PnP.Core.Test.Base
 {
