+++ src/sdk/PnP.Core/Model/Teams/Internal/TeamAppCollection.cs
@@ -1,6 +1,4 @@
-﻿using System.Linq;
-
-namespace PnP.Core.Model.Teams
+﻿namespace PnP.Core.Model.Teams
 {
     internal partial class TeamAppCollection
     {
