+++ src/sdk/PnP.Core/Model/Teams/Internal/TeamChannel.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using System.Dynamic;
 using System.Text.Json;
+using PnP.Core.Model.Teams.Public.Enums;
 using PnP.Core.Services;
 
 namespace PnP.Core.Model.Teams
