+++ src/sdk/PnP.Core/Model/Teams/Internal/TeamChannelCollection.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using System.Threading.Tasks;
 using PnP.Core.Services;
