+++ src/sdk/PnP.Core/Model/Teams/Internal/TeamChatMessageReaction.cs
@@ -4,17 +4,15 @@ using PnP.Core.Services;
 
 namespace PnP.Core.Model.Teams
 {
-    [GraphType(Uri = "teams/{Parent.GraphId}/installedapps/{GraphId}")]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "CA2243:Attribute string literals should parse correctly", Justification = "<Pending>")]
-    internal partial class TeamApp
+    internal partial class TeamChatMessageReaction
     {
-        public TeamApp()
+        public TeamChatMessageReaction()
         {
             MappingHandler = (FromJson input) =>
             {
                 switch (input.TargetType.Name)
                 {
-                    case "TeamsAppDistributionMethod": return JsonMappingHelper.ToEnum<TeamsAppDistributionMethod>(input.JsonElement);
+                    case "ChatMessageReactionType": return JsonMappingHelper.ToEnum<ChatMessageReactionType>(input.JsonElement);
                 }
 
                 input.Log.LogDebug($"Field {input.FieldName} could not be mapped when converting from JSON");
