+++ src/sdk/PnP.Core/Model/Teams/Internal/TeamChatMessageReaction.gen.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using PnP.Core.Model.Teams.Public.Enums;
 
 namespace PnP.Core.Model.Teams
 {
@@ -7,7 +8,7 @@ namespace PnP.Core.Model.Teams
     {
         public DateTimeOffset CreatedDateTime { get => GetValue<DateTimeOffset>(); set => SetValue(value); }
 
-        public string ReactionType { get => GetValue<string>(); set => SetValue(value); }
+        public ChatMessageReactionType ReactionType { get => GetValue<ChatMessageReactionType>(); set => SetValue(value); }
 
         public ITeamIdentitySet User { get => GetValue<ITeamIdentitySet>(); set => SetValue(value); }
     }
