+++ src/sdk/PnP.Core/Model/Teams/Internal/TeamFunSettings.gen.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
+using PnP.Core.Model.Teams.Public.Enums;
 using PnP.Core.Services;
 
 namespace PnP.Core.Model.Teams
