+++ src/sdk/PnP.Core/Model/Teams/Public/Enums/ChatMessageReactionType.cs
@@ -1,0 +1,15 @@
+﻿namespace PnP.Core.Model.Teams.Public.Enums
+{
+    /// <summary>
+    /// Types of reactions to a chat message
+    /// </summary>
+    public enum ChatMessageReactionType
+    {
+        Like,
+        Angry,
+        Sad,
+        Laugh,
+        Heart,
+        Surprised
+    }
+}
