+++ src/sdk/PnP.Core/Model/Teams/Public/Enums/TeamChannelMembershipType.cs
@@ -1,0 +1,12 @@
+﻿namespace PnP.Core.Model.Teams.Public.Enums
+{
+    /// <summary>
+    /// Defines the Membership Type for a Team Channel
+    /// </summary>
+    public enum TeamChannelMembershipType
+    {
+        Standard,
+        Private,
+        UnknownFutureValue,
+    }
+}
