+++ src/sdk/PnP.Core/Model/Teams/Public/Enums/TeamGiphyContentRating.cs
@@ -1,0 +1,11 @@
+﻿namespace PnP.Core.Model.Teams.Public.Enums
+{
+    /// <summary>
+    /// Giphy content rating for giphies being used in a team
+    /// </summary>
+    public enum TeamGiphyContentRating
+    {
+        Moderate,
+        Strict
+    }
+}
