+++ src/sdk/PnP.Core/Model/Teams/Public/Enums/TeamsAppDistributionMethod.cs
@@ -1,0 +1,21 @@
+﻿namespace PnP.Core.Model.Teams.Public.Enums
+{
+    /// <summary>
+    /// Distribution method of the team app
+    /// </summary>
+    public enum TeamsAppDistributionMethod
+    {
+        /// <summary>
+        /// From store
+        /// </summary>
+        Store,
+        /// <summary>
+        /// From your organization
+        /// </summary>
+        Organization,
+        /// <summary>
+        /// Side loaded (developer scenario)
+        /// </summary>
+        SideLoaded
+    }
+}
