+++ src/sdk/PnP.Core/Model/Teams/Public/ITeamApp.cs
@@ -1,4 +1,6 @@
-﻿namespace PnP.Core.Model.Teams
+﻿using PnP.Core.Model.Teams.Public.Enums;
+
+namespace PnP.Core.Model.Teams
 {
     /// <summary>
     /// Defines a Microsoft Teams App installation
@@ -27,24 +29,5 @@
         public TeamsAppDistributionMethod DistributionMethod { get; set; }
 
         // Note: we intentionally left out the collection of TeamsAppDefinition from this prototype
-    }
-
-    /// <summary>
-    /// Distribution method of the team app
-    /// </summary>
-    public enum TeamsAppDistributionMethod
-    {
-        /// <summary>
-        /// From store
-        /// </summary>
-        Store,
-        /// <summary>
-        /// From your organization
-        /// </summary>
-        Organization,
-        /// <summary>
-        /// Side loaded (developer scenario)
-        /// </summary>
-        Sideloaded
     }
 }
