+++ src/sdk/PnP.Core/Model/Teams/Public/ITeamAsyncOperation.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 
 namespace PnP.Core.Model.Teams
 {
